Index chat page
This is a testing version that uses the partial file _Conversations to create a container in the left part of the screen with all the conversations (there are testing convo for now) and in the right part there is a ChatRoom container that will include message history(look _ChatRoom partial file).
The index method of chat controller , fetches the conversations and the RenderChat will retrieve messages from the database.
Chat hub contains 2 different send message methods.
fetch_chat.js is used to call the renderChat method